requirements.txt
Django
django-taggit
django-crispy-forms
crispy-bootstrap5
django-recaptcha